+++ CPayDemoOC/Pods/Target Support Files/CPaySDK/CPaySDK-xcframeworks.sh
@@ -23,12 +23,6 @@ variant_for_slice()
   "CPaySDK.xcframework/ios-arm64_x86_64-simulator")
     echo "simulator"
     ;;
-  "PPRiskMagnes.xcframework/ios-arm64")
-    echo ""
-    ;;
-  "PPRiskMagnes.xcframework/ios-arm64_x86_64-simulator")
-    echo "simulator"
-    ;;
   "PayKitUI.xcframework/ios-arm64")
     echo ""
     ;;
@@ -39,36 +33,6 @@ variant_for_slice()
     echo ""
     ;;
   "PayKit.xcframework/ios-arm64_x86_64-simulator")
-    echo "simulator"
-    ;;
-  "KlarnaExt.xcframework/ios-arm64")
-    echo ""
-    ;;
-  "KlarnaExt.xcframework/ios-arm64_x86_64-simulator")
-    echo "simulator"
-    ;;
-  "PaymentButtons.xcframework/ios-arm64")
-    echo ""
-    ;;
-  "PaymentButtons.xcframework/ios-arm64_x86_64-simulator")
-    echo "simulator"
-    ;;
-  "CorePayments.xcframework/ios-arm64")
-    echo ""
-    ;;
-  "CorePayments.xcframework/ios-arm64_x86_64-simulator")
-    echo "simulator"
-    ;;
-  "FraudProtection.xcframework/ios-arm64")
-    echo ""
-    ;;
-  "FraudProtection.xcframework/ios-arm64_x86_64-simulator")
-    echo "simulator"
-    ;;
-  "PayPalWebPayments.xcframework/ios-arm64")
-    echo ""
-    ;;
-  "PayPalWebPayments.xcframework/ios-arm64_x86_64-simulator")
     echo "simulator"
     ;;
   esac
@@ -83,12 +47,6 @@ archs_for_slice()
   "CPaySDK.xcframework/ios-arm64_x86_64-simulator")
     echo "arm64 x86_64"
     ;;
-  "PPRiskMagnes.xcframework/ios-arm64")
-    echo "arm64"
-    ;;
-  "PPRiskMagnes.xcframework/ios-arm64_x86_64-simulator")
-    echo "arm64 x86_64"
-    ;;
   "PayKitUI.xcframework/ios-arm64")
     echo "arm64"
     ;;
@@ -99,36 +57,6 @@ archs_for_slice()
     echo "arm64"
     ;;
   "PayKit.xcframework/ios-arm64_x86_64-simulator")
-    echo "arm64 x86_64"
-    ;;
-  "KlarnaExt.xcframework/ios-arm64")
-    echo "arm64"
-    ;;
-  "KlarnaExt.xcframework/ios-arm64_x86_64-simulator")
-    echo "arm64 x86_64"
-    ;;
-  "PaymentButtons.xcframework/ios-arm64")
-    echo "arm64"
-    ;;
-  "PaymentButtons.xcframework/ios-arm64_x86_64-simulator")
-    echo "arm64 x86_64"
-    ;;
-  "CorePayments.xcframework/ios-arm64")
-    echo "arm64"
-    ;;
-  "CorePayments.xcframework/ios-arm64_x86_64-simulator")
-    echo "arm64 x86_64"
-    ;;
-  "FraudProtection.xcframework/ios-arm64")
-    echo "arm64"
-    ;;
-  "FraudProtection.xcframework/ios-arm64_x86_64-simulator")
-    echo "arm64 x86_64"
-    ;;
-  "PayPalWebPayments.xcframework/ios-arm64")
-    echo "arm64"
-    ;;
-  "PayPalWebPayments.xcframework/ios-arm64_x86_64-simulator")
     echo "arm64 x86_64"
     ;;
   esac
@@ -213,13 +141,7 @@ install_xcframework() {
   echo "Copied $source to $destination"
 }
 
-install_xcframework "${PODS_ROOT}/CPaySDK/CPaySDK/Core/CPaySDK.xcframework" "CPaySDK/Core" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
-install_xcframework "${PODS_ROOT}/CPaySDK/CPaySDK/Ext/PPRiskMagnes.xcframework" "CPaySDK/Ext/PPRiskMagnes" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
-install_xcframework "${PODS_ROOT}/CPaySDK/CPaySDK/Payment/CashApp/Ext/PayKitUI.xcframework" "CPaySDK/Payment/CashApp/Button" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
-install_xcframework "${PODS_ROOT}/CPaySDK/CPaySDK/Payment/CashApp/Ext/PayKit.xcframework" "CPaySDK/Payment/CashApp/Core" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
-install_xcframework "${PODS_ROOT}/CPaySDK/CPaySDK/Payment/Klarna/KlarnaExt.xcframework" "CPaySDK/Payment/Klarna" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
-install_xcframework "${PODS_ROOT}/CPaySDK/CPaySDK/Payment/PayPal/Ext/PaymentButtons.xcframework" "CPaySDK/Payment/PayPal/Button" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
-install_xcframework "${PODS_ROOT}/CPaySDK/CPaySDK/Payment/PayPal/Ext/CorePayments.xcframework" "CPaySDK/Payment/PayPal/Core" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
-install_xcframework "${PODS_ROOT}/CPaySDK/CPaySDK/Payment/PayPal/Ext/FraudProtection.xcframework" "CPaySDK/Payment/PayPal/Fraud" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
-install_xcframework "${PODS_ROOT}/CPaySDK/CPaySDK/Payment/PayPal/Ext/PayPalWebPayments.xcframework" "CPaySDK/Payment/PayPal/Web" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
+install_xcframework "${PODS_ROOT}/../../CPaySDK/Core/CPaySDK.xcframework" "CPaySDK/Core" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
+install_xcframework "${PODS_ROOT}/../../CPaySDK/Payment/CashApp/Ext/PayKitUI.xcframework" "CPaySDK/Payment/CashApp/Button" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"
+install_xcframework "${PODS_ROOT}/../../CPaySDK/Payment/CashApp/Ext/PayKit.xcframework" "CPaySDK/Payment/CashApp/Core" "framework" "ios-arm64" "ios-arm64_x86_64-simulator"